Use statement class and reorg code
