+++ src/index.js
@@ -1,5 +1,4 @@
-const getQueriesTokens = require("./get-queries-tokens");
-const enforceLimit = require("./enforce-limit.js");
+const getStatements = require("./get-statements");
 
 const VALID_STRATEGIES = ["limit", "fetch"];
 
@@ -40,32 +39,24 @@ function limit(sqlText, limitStrategies, limitNumber) {
     }
   });
 
-  let enforcedSql = "";
-
-  const queriesTokens = getQueriesTokens(sqlText);
-
-  queriesTokens.forEach((queryTokens) => {
-    let enforcedTokens = [];
-    enforcedTokens = enforceLimit(queryTokens, strategies, limitNumber);
-    enforcedSql += enforcedTokens.map((t) => t.text).join("");
-  });
-
-  // Return a string
-  return enforcedSql;
+  return getStatements(sqlText)
+    .map((statement) => {
+      statement.enforceLimit(strategies, limitNumber);
+      return statement.toString();
+    })
+    .join("");
 }
 
 /**
  * Splits SQL text on terminator, returning an array of SQL statements
  * @param {string} sqlText
  */
-function getStatements(sqlText) {
+function apiGetStatements(sqlText) {
   if (typeof sqlText !== "string") {
     throw new Error("sqlText must be string");
   }
-  const queriesTokens = getQueriesTokens(sqlText);
-  return queriesTokens.map((queryTokens) =>
-    queryTokens.map((t) => t.text).join("")
-  );
+  const statements = getStatements(sqlText);
+  return statements.map((statement) => statement.toString());
 }
 
 /**
@@ -78,16 +69,9 @@ function removeTerminator(sqlStatement) {
   if (typeof sqlStatement !== "string") {
     throw new Error("sqlText must be string");
   }
-  const queriesTokens = getQueriesTokens(sqlStatement);
-
-  const statements = queriesTokens
-    .map((queryTokens) =>
-      queryTokens
-        .filter((t) => t.type !== "terminator")
-        .map((t) => t.text)
-        .join("")
-    )
-    .filter((statement) => statement.trim() !== "");
+  const statements = getStatements(sqlStatement)
+    .map((s) => s.toString(true))
+    .filter((s) => s.trim() !== "");
 
   if (statements.length > 1) {
     throw new Error("Multiple statements detected");
@@ -98,6 +82,6 @@ function removeTerminator(sqlStatement) {
 
 module.exports = {
   limit,
-  getStatements,
+  getStatements: apiGetStatements,
   removeTerminator,
 };
